feat(ui): add global chat page at /chat
Extract ChatPanel from session chat page into a shared component
so both session chat and the new global chat reuse the same UI.

- New /chat route with lazy session creation (one per org, Generic harness)
- ChatPanel component at components/chat/chat-panel.tsx
- useGlobalChat hook manages session persistence via localStorage
- "Chat" added to sidebar above Dashboard
