+++ apps/ui/src/hooks/index.ts
@@ -9,3 +9,4 @@ export * from "./use-auth";
 export * from "./use-durable";
 export * from "./use-image-attachments";
 export * from "./use-session-schedules";
+export * from "./use-global-chat";
